requirements.txt
streamlit>=1.46.0
rasterio>=1.3.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
geopandas>=1.0.0
matplotlib>=3.7.0
pandas>=2.0.0
shapely>=2.0.0
streamlit-image-coordinates>=0.1.6
torch>=2.0.0
torchvision>=0.15.0